refactor: 레포지토리 변경을 위해 interface 사용
